% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{swift_pal}
\alias{swift_pal}
\title{Function to interpolate a color palette}
\usage{
swift_pal(palette = "taylor1989", reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in tswift_palettes}

\item{reverse}{Boolean true if palette should be reversed}

\item{...}{Additional arguments to pass to colorRampPalette()}
}
\value{
A vector of colors
}
\description{
Function to interpolate a color palette
}
