% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{swift_palettes}
\alias{swift_palettes}
\title{Color palettes}
\format{
An object of class \code{list} of length 11.
}
\usage{
swift_palettes
}
\description{
Color palettes
}
\keyword{datasets}
