% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{scale_fill_taylor}
\alias{scale_fill_taylor}
\title{Fill scale for Taylor Swift colors}
\usage{
scale_fill_taylor(
  palette = "taylor1989",
  discrete = TRUE,
  reverse = FALSE,
  ...
)
}
\arguments{
\item{palette}{Character name of palette in swift_palettes}

\item{discrete}{Boolean if color aesthetic is discrete}

\item{reverse}{Boolean if palette should be reversed}

\item{...}{Additional arguments used to discrete_scale() or scale_fill_gradientn()
to automatically interpolate between colours.}
}
\value{
No return value. Called for side effects
}
\description{
Fill scale for Taylor Swift colors
}
\examples{
library(ggplot2)
data <- data.frame(c = LETTERS[1:3],x = c(1,5,7),y = c(5,9,13))
ggplot(data, aes(x,fill=c))+geom_bar()+scale_fill_taylor()
}
